Move default template from agent pattern to route-only
Replace the createAgent() indirection with inline route handlers so the
default template is simpler and easier to follow. Translation logic now
lives directly in the API routes using c.var.* for services.

- Remove src/agent/translate/ (agent + evals)
- Inline translation logic into POST /translate route handler
- Clean up base template's phantom @agent/hello import
- Drop @agentuity/evals and groq-sdk dependencies
- Update App.tsx next-steps copy to reference routes
