+++ templates/_base/src/api/index.ts
@@ -1,11 +1,8 @@
 import { Hono } from 'hono';
 import type { Env } from '@agentuity/runtime';
-import hello from '@agent/hello';
 
-const api = new Hono<Env>().post('/hello', hello.validator(), async (c) => {
-	const data = c.req.valid('json');
-	const result = await hello.run(data);
-	return c.json(result);
+const api = new Hono<Env>().get('/hello', (c) => {
+	return c.json({ message: 'Hello from Agentuity!' });
 });
 
 export type ApiRouter = typeof api;
