--- templates/default/src/agent/index.ts
@@ -1,7 +1,0 @@
-/**
- * Agent registry — import your agents here and export them as an array.
- * The array is passed to createApp() in app.ts to register all agents.
- */
-import translate from './translate';
-
-export default [translate];
